+++ src/Components/Attendees.js
@@ -1,3 +1,31 @@
-export default function Attendees() {
-  return;
+import Attendee from "./Attendee";
+import { useState } from "react";
+
+export default function Attendees( {attendees},{updateEventAttendance}) { //, {attendees}, {setEvents}
+  console.log("hello")
+  const [showAttendees, setShowAttendees] = useState(false);
+  
+
+  function toggleEventAttendees() {
+    setShowAttendees(!showAttendees);
+  }
+
+  
+ 
+  return (
+    <>
+    <button onClick={toggleEventAttendees}>
+        {!showAttendees ? "Show Attendees" : "Hide Attendees"}
+      </button>
+      {showAttendees ? (
+        <div className="attendees">
+        {attendees.map((attendee, index) => (
+        <Attendee attendee={attendee}  updateEventAttendance={updateEventAttendance} />
+        ))}
+        </div>
+      ) : null}
+    </>
+  )
 }
+
+
